services/ai-worker/requirements.txt
kafka-python==2.0.2
aiokafka==0.8.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
google-generativeai==0.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
uvloop==0.19.0
python-json-logger==2.0.7
